+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v20.0.0](https://github.com/lexicalunit/spellbot/releases/tag/v20.0.0) - 2026-05-25
+
 ### Changed
 
 - Renamed `/score` to `/record` since we don't track points or elo anymore.